move tesseract exported wasm back into tesseract/mod.rs
